requirements.txt
anyjson==0.3.3
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.16.1
cffi==1.8.3
channels==0.17.3
circus==0.14.0
circus-web==1.0.0
constantly==15.1.0
cryptography==1.5.2
daphne==0.15.0
Django==1.10.3
django-debug-toolbar==1.6
django-eventtools==0.9.10
django-redis==4.6.0
httplib2==0.9.2
idna==2.1
incremental==16.10.1
iowait==0.2
Mako==1.0.5
MarkupSafe==0.23
msgpack-python==0.4.8
psutil==4.4.2
psycopg2==2.6.2
pyasn1==0.1.9
pycparser==2.17
PySocks==1.5.7
python-dateutil==2.5.3
pytz==2016.7
pyzmq==16.0.0
redis==2.10.5
requests==2.11.1
schedule==0.3.2
simplejson==3.10.0
six==1.10.0
slacker==0.9.29
sqlparse==0.2.2
tomako==0.1.0.post1
TornadIO2==0.0.3
tornado==4.4.2
twilio==5.6.0
Twisted==16.5.0
txaio==2.5.2
urllib3==1.19
zope.interface==4.3.2
numpy==1.11.3
pandas==0.19.1
